requirements.txt
tensorflow
numpy
opencv-python
Pillow
streamlit
scikit-learn
